refactor: mfsdk
